Unit test the file parsing function.
I wonder if this is the right level to test.  I could test at lower levels,
but … I mean, at the end of the day, I only care that the buffer contents turn
into a specific hash table format.  I don’t particularly care what happens in
the code to make that happen.  So maybe that makes this the right level to test?